policyme-graphrag-engine/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart
neo4j==6.1.0
google-genai
langchain-text-splitters>=0.2.0
aiokafka==0.13.0
requests==2.32.5
numpy==2.4.2
